apps/ingest-service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
requests==2.32.3
python-barcode==0.15.1
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
